+++ interface/src/examples/serial/SerialInfoWithMode.tsx
@@ -1,0 +1,15 @@
+import React, { FC } from 'react';
+import { Box } from '@mui/material';
+import { UartModeSwitcher } from '../../components';
+import SerialInfo from './SerialInfo';
+
+const SerialInfoWithMode: FC = () => {
+  return (
+    <Box>
+      <UartModeSwitcher currentMode="live" />
+      <SerialInfo />
+    </Box>
+  );
+};
+
+export default SerialInfoWithMode;