+++ interface/src/examples/serial/SerialMonitor.tsx
@@ -3,7 +3,7 @@ import { Navigate, Route, Routes } from 'react-router-dom';
 import { Tab } from '@mui/material';
 import { RouterTabs, useRouterTab, useLayoutTitle } from '../../components';
 
-import SerialInfo from './SerialInfo';
+import SerialInfoWithMode from './SerialInfoWithMode';
 import SerialConfig from './SerialConfig';
 import SerialRest from './SerialRest';
 import SerialWebSocket from './SerialWebSocket';
@@ -23,7 +23,7 @@ const SerialMonitor: FC = () => {
         <Tab value="ble" label="BLE Stream" />
       </RouterTabs>
       <Routes>
-        <Route path="information" element={<SerialInfo />} />
+        <Route path="information" element={<SerialInfoWithMode />} />
         <Route path="configuration" element={<SerialConfig />} />
         <Route path="rest" element={<SerialRest />} />
         <Route path="stream" element={<SerialWebSocket />} />